# backend/db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.25.0

package db

import (
	"database/sql"
	"time"

	"github.com/google/uuid"
)

type Chats struct {
	ID          uuid.UUID `json:"id"`
	Content     string    `json:"content"`
	CreatedAt   time.Time `json:"created_at"`
	SenderUid   uuid.UUID `json:"sender_uid"`
	ReceiverUid uuid.UUID `json:"receiver_uid"`
}

type Users struct {
	Uid       uuid.UUID      `json:"uid"`
	Name      sql.NullString `json:"name"`
	Email     string         `json:"email"`
	Avatar    sql.NullString `json:"avatar"`
	Password  sql.NullString `json:"password"`
	Provider  string         `json:"provider"`
	CreatedAt time.Time      `json:"created_at"`
}